+++ pkg/config/aqua/config.go
@@ -48,20 +48,6 @@ type Config struct {
 	Checksum   *Checksum  `json:"checksum"`
 }
 
-func (cfg *Config) ChecksumEnabled() bool {
-	if cfg == nil {
-		return false
-	}
-	if cfg.Checksum == nil {
-		return false
-	}
-	return cfg.Checksum.Enabled
-}
-
-type Checksum struct {
-	Enabled bool `json:"enabled"`
-}
-
 type Registries map[string]*Registry
 
 func (Registries) JSONSchema() *jsonschema.Schema {
